fix: Refactor tools module
