+++ android/gradle.properties
@@ -1,4 +1,3 @@
 org.gradle.jvmargs=-Xmx1536M
 android.useAndroidX=true
 android.enableJetifier=true
-org.gradle.java.home=D:\\Android\\Android Studio\\jre
