+++ lib/app/modules/main/submodules/tools/plant_diseases/recognition_interface.dart
@@ -1,0 +1,7 @@
+import 'package:image/image.dart';
+
+abstract class IRecognition {
+  Future<void> loadModel();
+  Future<void> loadLabels();
+  dynamic predict(Image image);
+}
