+++ lib/app/modules/main/submodules/tools/plant_diseases/unsuported_platform.dart
@@ -1,0 +1,7 @@
+import 'recognition_interface.dart';
+
+abstract class Classifier implements IRecognition {
+  Classifier() {
+    throw UnimplementedError("Plataforma não suportada");
+  }
+}
